Move classes in main postgis-jdbc module to net.postgis.jdbc
