+++ jdbc/src/main/java/net/postgis/jdbc/PGbox2d.java
@@ -25,7 +25,9 @@
  * 
  */
 
-package net.postgis;
+package net.postgis.jdbc;
+
+import net.postgis.Point;
 
 import java.sql.SQLException;
