+++ jdbc/src/main/java/net/postgis/jdbc/PGgeographyLW.java
@@ -13,44 +13,48 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  *
- * (C) 2004 Paul Ramsey, pramsey@refractions.net
- *
  * (C) 2005 Markus Schaber, markus.schaber@logix-tt.com
  *
  * (C) 2015 Phillip Ross, phillip.w.g.ross@gmail.com
  */
 
-package net.postgis;
+package net.postgis.jdbc;
 
+
+import net.postgis.Geometry;
+import net.postgis.binary.BinaryWriter;
 
 import java.sql.SQLException;
 
 
 /**
- * A PostgreSQL JDBC PGobject extension data type modeling the geometry type.
+ * A PostgreSQL JDBC PGobject extension data type modeling the geography type.
+ *
+ * The hex-encoded EWKB format is used to communicate with the backend, which is much more efficient,
+ * but only works with Lwgeom enabled PostGIS (1.0.0 and up).
  *
  * @author Phillip Ross
  */
-public class PGgeometry extends PGgeo {
+public class PGgeographyLW extends PGgeography {
 
-    private static final long serialVersionUID = 4116907189503026815L;
+    private static final long serialVersionUID = 7717856818804158022L;
+
+    /** The binary writer to be used for serializing geometry to a PGobject value. */
+    BinaryWriter bw = new BinaryWriter();
 
 
     /** Instantiate with default state. */
-    public PGgeometry() {
+    public PGgeographyLW() {
         super();
-        setType("geometry");
     }
-
 
     /**
      * Instantiate with the specified state.
      *
      * @param geometry the geometry to instantiate with
      */
-    public PGgeometry(final Geometry geometry) {
+    public PGgeographyLW(final Geometry geometry) {
         super(geometry);
-        setType("geometry");
     }
 
 
@@ -59,16 +63,22 @@ public class PGgeometry extends PGgeo {
      *
      * @param value the value to instantiate with
      */
-    public PGgeometry(final String value) throws SQLException {
+    public PGgeographyLW(final String value) throws SQLException {
         super(value);
-        setType("geometry");
+    }
+
+
+    /** {@inheritDoc} */
+    @Override
+    public String getValue() {
+        return bw.writeHexed(geometry);
     }
 
 
     /** {@inheritDoc} */
     @Override
     public Object clone() {
-        return new PGgeometry(geometry);
+        return new PGgeographyLW(geometry);
     }
