+++ jdbc/src/main/resources/org/postgresql/driverconfig.properties
@@ -3,7 +3,7 @@
 # PostGIS datatypes within the jdbc driver.
 #
 
-datatype.geometry=net.postgis.PGgeometry
-datatype.geography=net.postgis.PGgeography
-datatype.box3d=net.postgis.PGbox3d
-datatype.box2d=net.postgis.PGbox2d
+datatype.geometry=net.postgis.jdbc.PGgeometry
+datatype.geography=net.postgis.jdbc.PGgeography
+datatype.box3d=net.postgis.jdbc.PGbox3d
+datatype.box2d=net.postgis.jdbc.PGbox2d
