+++ jdbc/src/test/java/net/postgis/jdbc/DatatypesAutoRegistrationTest.java
@@ -18,7 +18,7 @@
  * (C) 2020 Phillip Ross, phillip.w.g.ross@gmail.com
  */
 
-package net.postgis;
+package net.postgis.jdbc;
 
 
 import net.postgis.tools.testutils.TestContainerController;
