+++ jdbc/src/test/java/net/postgis/jdbc/SerializationTest.java
@@ -1,14 +1,13 @@
 /*
- * PGbox3d.java
- * 
- * PostGIS extension for PostgreSQL JDBC driver - bounding box model
- * 
- * 
+ * ServerTest.java
+ *
+ * PostGIS extension for PostgreSQL JDBC driver - example and test classes
+ *
  * (C) 2004 Paul Ramsey, pramsey@refractions.net
- * 
+ *
  * (C) 2005 Markus Schaber, markus.schaber@logix-tt.com
- * 
- * (C) 2015 Phillip Ross, phillip.w.g.ross@gmail.com
+ *
+ * (C) 2017 Phillip Ross, phillip.w.g.ross@gmail.com
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -23,38 +22,34 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- * 
+ *
  */
 
-package net.postgis;
+package net.postgis.jdbc;
 
-import java.sql.SQLException;
 
-public class PGbox3d extends PGboxbase {
-    /* JDK 1.5 Serialization */
-    private static final long serialVersionUID = 0x100;
+import net.postgis.jdbc.PGgeometry;
+import org.testng.Assert;
+import org.testng.annotations.Test;
 
-    public PGbox3d() {
-        super();
+import java.io.ByteArrayOutputStream;
+import java.io.NotSerializableException;
+import java.io.ObjectOutputStream;
+
+
+public class SerializationTest {
+
+
+    @Test
+    public void serializationCheckPGgeometry() throws Exception {
+        try {
+            new ObjectOutputStream(new ByteArrayOutputStream())
+                    .writeObject(new PGgeometry("MULTIPOLYGON(((1 1,1 2,2 1,1 1)))"));
+        }
+        catch (NotSerializableException ex) {
+            Assert.fail("serialization of PGgeometry failed: " + ex);
+        }
     }
 
-    public PGbox3d(Point llb, Point urt) {
-        super(llb, urt);
-    }
 
-    public PGbox3d(String value) throws SQLException {
-        super(value);
-    }
-
-    public String getPrefix() {
-        return ("BOX3D");
-    }
-
-    public String getPGtype() {
-        return ("box3d");
-    }
-
-    protected PGboxbase newInstance() {
-        return new PGbox3d();
-    }
 }
