+++ jdbc/src/test/java/net/postgis/jdbc/ServerTest.java
@@ -25,7 +25,7 @@
  * 
  */
 
-package net.postgis;
+package net.postgis.jdbc;
 
 
 import net.postgis.tools.testutils.TestContainerController;
