+++ jdbc/src/test/java/net/postgis/jdbc/util/VersionUtilIT.java
@@ -16,10 +16,12 @@
  * (C) 2018 Phillip Ross, phillip.w.g.ross@gmail.com
  */
 
-package net.postgis.util;
+package net.postgis.jdbc.util;
 
 
 import net.postgis.tools.testutils.TestContainerController;
+import net.postgis.util.VersionFunctions;
+import net.postgis.util.VersionUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testng.Assert;
